minor zsh_recompile update

+++ config/zsh/functions.zsh
@@ -14,15 +14,17 @@ function zsh_recompile {
     [[ -f $f ]] && zrecompile -p $f
     [[ -f $f.zwc.old ]] && rm -f $f.zwc.old
   done
-  zrecompile -p ~/.config/zsh/fasd/fasd
 
   [[ -f ~/.cache/zsh/compdump ]] && zrecompile -p ~/.cache/zsh/compdump
   [[ -f ~/.cache/zsh/compdump.zwc.old ]] && rm -f ~/.cache/zsh/compdump.zwc.old
 
-  [[ -f $FASD_INIT_CACHE.zwc ]] && rm -f $FASD_INIT_CACHE.zwc
-  [[ -f $FASD_INIT_CACHE ]] && rm -f $FASD_INIT_CACHE
-  fasd --init posix-alias zsh-hook zsh-ccomp zsh-ccomp-install zsh-wcomp zsh-wcomp-install > $FASD_INIT_CACHE
-  zrecompile -p $FASD_INIT_CACHE
+  zrecompile -p ~/.config/zsh/fasd/fasd
+  if [[ ! -z "$FASD_INIT_CACHE" ]]; then
+      [[ -f $FASD_INIT_CACHE.zwc ]] && rm -f $FASD_INIT_CACHE.zwc
+      [[ -f $FASD_INIT_CACHE ]] && rm -f $FASD_INIT_CACHE
+      fasd --init posix-alias zsh-hook zsh-ccomp zsh-ccomp-install zsh-wcomp zsh-wcomp-install > $FASD_INIT_CACHE
+      zrecompile -p $FASD_INIT_CACHE
+  fi
 
   source ~/.zshrc
 }
